chore: moving earned badges from local to gloabal components
