+++ components/EarnedBadges.tsx
